requirements.txt
discord.py
google-cloud-vision
google-oauth
asyncio
